Device autodetection via sysfs.
If support for sysfs is compiled in and the libsysfs library is found, the
acecad driver will now try to autodetect the device via sysfs at first. If this
fails, it will probe /dev/input/eventX devices.

Currently the Name option is ignored during sysfs detection: the first possibly
valid candidate is accepted without further checks.